Create KYC
Right Coin's proposal is that: What is Rightcoin? It is the union of several international demands in entertainment. Podcasts, youtubers, gamers, streaming, metaverse, news channels, music, culture and education gathered in a single platform that will unite the already established content producer, the content producer at the beginning of the activity to the final consumer.